feat(web): versão web inicial + guia de migração
- LoginScreen, HomeScreenWeb, GamesTab, GroupsTab (wasmJs)
- FirebaseManager mock (substituir por external declarations)
- AppNavigation expect/actual (multiplataforma)
- WEB_MIGRATION_GUIDE.md (6000+ palavras para LLMs)
- create-web-migration-issues.js (21 issues #163-#183)
- Hot reload localhost:8081

Próximo: Issue #163 (Firebase Auth real)

+++ app/src/main/java/com/futebadosparcas/domain/ranking/LeagueService.kt
@@ -263,7 +263,7 @@ class LeagueService constructor(
 
              if (lastParticipation != null) {
                 // Bug #2 Fix: Busca ultima participacao independente do mes
-                val sharedDivision = com.futebadosparcas.domain.ranking.LeagueRatingCalculator.getDivisionForRating(lastParticipation.leagueRating)
+                val sharedDivision = com.futebadosparcas.domain.ranking.LeagueRatingCalculator.getDivisionForRating(lastParticipation.leagueRating.toDouble())
                 startDivision = sharedDivision.name
 
                 // Bug #3 Fix: Momentum - FIXME: recentGames nao existe no modelo KMP
